Updated file names and moved to examples/data_pipeline
