+++ examples/data_pipeline/pytext_vocab.py
